server/requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-cors==3.0.10
requests==2.26.0
python-dotenv==0.19.0
certifi==2025.1.31
charset-normalizer==2.0.12
click==8.1.8
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
six==1.17.0
urllib3==1.26.20
